refactor(adapter-ui): extract proposal review outcome renderers into a module
proposal-review.tsx grew past the 500-line file-size policy after the
pre-analysis render block was added. Move the two purely-presentational
outcome renderers (GraduateOutcome, MaterializeOutcome) into a sibling
module proposal-review-outcomes.tsx. No behavior change.

- proposal-review.tsx: 677 -> 471 lines
- proposal-review-outcomes.tsx: 211 lines (new)

Addresses CodeRabbit file-size review on #506.
