+++ addons/adapter-ui/cap-adapter-ui/src/pages/proposal-review.tsx
@@ -22,21 +22,19 @@ import {
 } from "@linchkit/ui-kit/components";
 import {
   AlertTriangleIcon,
-  CheckCircle2Icon,
   ClipboardListIcon,
   Code2Icon,
-  ExternalLinkIcon,
   GitPullRequestIcon,
   Loader2Icon,
   RefreshCwIcon,
   SparklesIcon,
   ThumbsDownIcon,
   ThumbsUpIcon,
-  XCircleIcon,
 } from "lucide-react";
 import { useCallback, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { ProposalImpactPreview } from "@/components/proposal-impact-preview";
+import { GraduateOutcome, MaterializeOutcome } from "@/components/proposal-review-outcomes";
 import {
   approveProposal,
   fetchProposals,
@@ -385,213 +383,6 @@ function ProposalCard({
       </CardContent>
     </Card>
   );
-}
-
-// ── Graduate outcome renderer ─────────────────────────────
-
-function GraduateOutcome({
-  result,
-  t,
-}: {
-  result: GraduateProposalResult;
-  t: ReturnType<typeof useTranslation>["t"];
-}) {
-  switch (result.kind) {
-    case "ok":
-      return (
-        <div
-          className="flex items-start gap-2 rounded-md border border-green-200 bg-green-50 p-3 dark:border-green-900 dark:bg-green-950/30"
-          data-testid="graduate-ok"
-        >
-          <CheckCircle2Icon className="mt-0.5 size-4 shrink-0 text-green-500" />
-          <div className="space-y-1">
-            <p className="text-sm text-green-700 dark:text-green-300">
-              {t("proposals.graduatePrOpened", "Pull request opened for review.")}
-            </p>
-            {result.prUrl && (
-              <a
-                href={result.prUrl}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="inline-flex items-center gap-1 text-xs text-primary hover:underline"
-              >
-                <ExternalLinkIcon className="size-3" />
-                {t("proposals.viewPr", "View PR")}
-              </a>
-            )}
-          </div>
-        </div>
-      );
-
-    case "unavailable":
-      return (
-        <div
-          className="flex items-start gap-2 rounded-md border border-dashed bg-muted/40 p-3"
-          data-testid="graduate-unavailable"
-        >
-          <AlertTriangleIcon className="mt-0.5 size-4 shrink-0 text-muted-foreground" />
-          <p className="text-sm text-muted-foreground">
-            {result.message ?? t("proposals.graduateNotConfigured", "Graduation not configured.")}
-          </p>
-        </div>
-      );
-
-    case "denied":
-      return (
-        <div
-          className="flex items-start gap-2 rounded-md bg-destructive/10 p-3"
-          data-testid="graduate-denied"
-        >
-          <XCircleIcon className="mt-0.5 size-4 shrink-0 text-destructive" />
-          <p className="text-sm text-destructive">
-            {t("proposals.notAuthorized", "Not authorized.")}
-          </p>
-        </div>
-      );
-
-    case "not_approved":
-      return (
-        <div
-          className="flex items-start gap-2 rounded-md bg-destructive/10 p-3"
-          data-testid="graduate-not-approved"
-        >
-          <AlertTriangleIcon className="mt-0.5 size-4 shrink-0 text-destructive" />
-          <p className="text-sm text-destructive">
-            {result.message ?? t("proposals.notApproved", "Proposal is not approved.")}
-          </p>
-        </div>
-      );
-
-    case "not_found":
-      return (
-        <div
-          className="flex items-start gap-2 rounded-md bg-destructive/10 p-3"
-          data-testid="graduate-not-found"
-        >
-          <AlertTriangleIcon className="mt-0.5 size-4 shrink-0 text-destructive" />
-          <p className="text-sm text-destructive">
-            {t("proposals.notFound", "Proposal not found.")}
-          </p>
-        </div>
-      );
-
-    case "error":
-      return (
-        <div
-          className="flex items-start gap-2 rounded-md bg-destructive/10 p-3"
-          data-testid="graduate-error"
-        >
-          <AlertTriangleIcon className="mt-0.5 size-4 shrink-0 text-destructive" />
-          <p className="text-sm text-destructive">{result.message}</p>
-        </div>
-      );
-  }
-}
-
-// ── Materialize outcome renderer ──────────────────────────
-
-function MaterializeOutcome({
-  result,
-  t,
-}: {
-  result: MaterializeProposalResult;
-  t: ReturnType<typeof useTranslation>["t"];
-}) {
-  switch (result.kind) {
-    case "ok": {
-      const materialized = result.outcomes.filter((o) => o.status === "materialized").length;
-      const skipped = result.outcomes.filter((o) => o.status === "skipped").length;
-      const failed = result.outcomes.filter((o) => o.status === "failed").length;
-      const ok = result.allMaterialized;
-      return (
-        <div
-          className={`flex items-start gap-2 rounded-md border p-3 ${
-            ok
-              ? "border-green-200 bg-green-50 dark:border-green-900 dark:bg-green-950/30"
-              : "border-amber-200 bg-amber-50 dark:border-amber-900 dark:bg-amber-950/30"
-          }`}
-          data-testid="materialize-ok"
-        >
-          {ok ? (
-            <CheckCircle2Icon className="mt-0.5 size-4 shrink-0 text-green-500" />
-          ) : (
-            <AlertTriangleIcon className="mt-0.5 size-4 shrink-0 text-amber-500" />
-          )}
-          <p className="text-sm">
-            {t("proposals.materializeSummary", "Generated {{m}}, skipped {{s}}, failed {{f}}.", {
-              m: materialized,
-              s: skipped,
-              f: failed,
-            })}
-          </p>
-        </div>
-      );
-    }
-
-    case "unavailable":
-      return (
-        <div
-          className="flex items-start gap-2 rounded-md border border-dashed bg-muted/40 p-3"
-          data-testid="materialize-unavailable"
-        >
-          <AlertTriangleIcon className="mt-0.5 size-4 shrink-0 text-muted-foreground" />
-          <p className="text-sm text-muted-foreground">
-            {result.message ??
-              t("proposals.materializeNotConfigured", "AI code generation is not configured.")}
-          </p>
-        </div>
-      );
-
-    case "denied":
-      return (
-        <div
-          className="flex items-start gap-2 rounded-md bg-destructive/10 p-3"
-          data-testid="materialize-denied"
-        >
-          <XCircleIcon className="mt-0.5 size-4 shrink-0 text-destructive" />
-          <p className="text-sm text-destructive">
-            {t("proposals.notAuthorized", "Not authorized.")}
-          </p>
-        </div>
-      );
-
-    case "not_draft":
-      return (
-        <div
-          className="flex items-start gap-2 rounded-md bg-destructive/10 p-3"
-          data-testid="materialize-not-draft"
-        >
-          <AlertTriangleIcon className="mt-0.5 size-4 shrink-0 text-destructive" />
-          <p className="text-sm text-destructive">
-            {result.message ?? t("proposals.notDraft", "Proposal is not a draft.")}
-          </p>
-        </div>
-      );
-
-    case "not_found":
-      return (
-        <div
-          className="flex items-start gap-2 rounded-md bg-destructive/10 p-3"
-          data-testid="materialize-not-found"
-        >
-          <AlertTriangleIcon className="mt-0.5 size-4 shrink-0 text-destructive" />
-          <p className="text-sm text-destructive">
-            {t("proposals.notFound", "Proposal not found.")}
-          </p>
-        </div>
-      );
-
-    case "error":
-      return (
-        <div
-          className="flex items-start gap-2 rounded-md bg-destructive/10 p-3"
-          data-testid="materialize-error"
-        >
-          <AlertTriangleIcon className="mt-0.5 size-4 shrink-0 text-destructive" />
-          <p className="text-sm text-destructive">{result.message}</p>
-        </div>
-      );
-  }
 }
 
 // ── Page ──────────────────────────────────────────────────
